requirements-api.txt
# Dépendances pour l'API FastAPI

# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Authentification JWT
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# CORS
python-dotenv==1.0.0

# Dépendances existantes (si nécessaire)
PyQt5==5.15.10
reportlab==4.0.7
